OS42, and cast changes based on @MJ666 suggestions

+++ src/main/drivers/pwm_output.c
@@ -126,7 +126,7 @@ static pwmOutputPort_t *pwmOutConfig(const timerHardware_t *timerHardware, uint8
 
 static void pwmWriteBrushed(uint8_t index, uint16_t value)
 {
-    *motors[index]->ccr = (uint16_t)((float)((value - 1000) * motors[index]->period / 1000));
+    *motors[index]->ccr = ((float)((value - 1000) * motors[index]->period / 1000));
 }
 
 static void pwmWriteStandard(uint8_t index, uint16_t value)
@@ -136,15 +136,15 @@ static void pwmWriteStandard(uint8_t index, uint16_t value)
 
 static void pwmWriteOneshot(uint8_t index, uint16_t value)
 {
-    *motors[index]->ccr = (uint16_t)((float)value);
+    *motors[index]->ccr = value;
 }
 
 static void pwmWriteMultiShot(uint8_t index, uint16_t value)
 {
 #ifdef STM32F4
-    *motors[index]->ccr = (uint16_t)((float)(value-1000) / 1.04166f)+ 240;
+    *motors[index]->ccr = ((float)(value-1000) / 1.04166f) + 240;
 #else
-    *motors[index]->ccr = (uint16_t)((float)(value-1000) / 2.08333f)+ 120;
+    *motors[index]->ccr = ((float)(value-1000) / 2.08333f) + 120;
 #endif
 }
